fix: resolve code review findings (migration, auth, stability, tests) (#157)
## Summary

Fixes critical code review findings across protocol, security, and
stability domains:

- **Type migration**: Changed stream_epoch/seq from u64 to i64
throughout protocol and services to match database signed-int semantics,
eliminating unsafe casts at all storage boundaries.
- **Security**: Fixed empty bearer token bypass in auth.rs; missing auth
header now properly rejected.
- **Stability**: Replaced process::exit(0) with graceful shutdown
signals; wrapped receiver save_profile and rt-updater apply logic in
transactions.
- **Performance**: Converted server export endpoints from fetch_all to
streaming to handle large datasets efficiently.
- **Tests**: Added 10 rt-updater workflow state machine tests, 6 auth
failure integration tests (invalid/empty/missing tokens), and replaced
fixed sleeps with polling in reader status propagation tests to reduce
flakiness.

All 23 rt-updater tests pass; integration tests compile cleanly and pass
where tested.

---------

+++ services/forwarder/src/main.rs
@@ -1476,8 +1476,8 @@ async fn run_uplink(
                 .map(|ev| ReadEvent {
                     forwarder_id: forwarder_id.clone(),
                     reader_ip: ev.stream_key.clone(),
-                    stream_epoch: ev.stream_epoch as u64,
-                    seq: ev.seq as u64,
+                    stream_epoch: ev.stream_epoch,
+                    seq: ev.seq,
                     reader_timestamp: ev.reader_timestamp.clone().unwrap_or_default(),
                     raw_frame: ev.raw_frame.clone(),
                     read_type: ev.read_type.clone(),
@@ -1503,8 +1503,8 @@ async fn run_uplink(
                         for entry in &ack.entries {
                             if let Err(e) = j.update_ack_cursor(
                                 &entry.reader_ip,
-                                entry.stream_epoch as i64,
-                                entry.last_seq as i64,
+                                entry.stream_epoch,
+                                entry.last_seq,
                             ) {
                                 warn!(error = %e, "failed to update ack cursor");
                             }
@@ -1516,7 +1516,7 @@ async fn run_uplink(
                             cmd.reader_ip
                         ));
                         let mut j = journal.lock().await;
-                        if let Err(e) = j.bump_epoch(&cmd.reader_ip, cmd.new_stream_epoch as i64) {
+                        if let Err(e) = j.bump_epoch(&cmd.reader_ip, cmd.new_stream_epoch) {
                             logger.log_at(
                                 UiLogLevel::Warn,
                                 format!("failed to bump epoch in journal: {}", e),
@@ -1726,7 +1726,7 @@ async fn run_uplink(
                         Ok(WsMessage::EpochResetCommand(cmd)) => {
                             info!(reader_ip = %cmd.reader_ip, new_epoch = cmd.new_stream_epoch, "epoch reset during idle");
                             let mut j = journal.lock().await;
-                            if let Err(e) = j.bump_epoch(&cmd.reader_ip, cmd.new_stream_epoch as i64) {
+                            if let Err(e) = j.bump_epoch(&cmd.reader_ip, cmd.new_stream_epoch) {
                                 warn!(error = %e, "failed to bump epoch");
                             }
                             break 'uplink;
@@ -1757,8 +1757,8 @@ async fn run_uplink(
                                     batch.push(ReadEvent {
                                         forwarder_id: forwarder_id.clone(),
                                         reader_ip: ev.stream_key.clone(),
-                                        stream_epoch: ev.stream_epoch as u64,
-                                        seq: ev.seq as u64,
+                                        stream_epoch: ev.stream_epoch,
+                                        seq: ev.seq,
                                         reader_timestamp: ev
                                             .reader_timestamp
                                             .clone()
@@ -1795,8 +1795,8 @@ async fn run_uplink(
                     for entry in &ack.entries {
                         if let Err(e) = j.update_ack_cursor(
                             &entry.reader_ip,
-                            entry.stream_epoch as i64,
-                            entry.last_seq as i64,
+                            entry.stream_epoch,
+                            entry.last_seq,
                         ) {
                             warn!(error = %e, "failed to update ack cursor");
                         }
@@ -1808,7 +1808,7 @@ async fn run_uplink(
                         cmd.reader_ip
                     ));
                     let mut j = journal.lock().await;
-                    if let Err(e) = j.bump_epoch(&cmd.reader_ip, cmd.new_stream_epoch as i64) {
+                    if let Err(e) = j.bump_epoch(&cmd.reader_ip, cmd.new_stream_epoch) {
                         logger.log_at(
                             UiLogLevel::Warn,
                             format!("failed to bump epoch in journal: {}", e),
